--- stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/builtin/systemlib/GetEnv.java
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://wso2.com) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.ballerinalang.nativeimpl.builtin.systemlib;
-
-import org.ballerinalang.bre.Context;
-import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BString;
-import org.ballerinalang.model.values.BValue;
-import org.ballerinalang.natives.AbstractNativeFunction;
-import org.ballerinalang.natives.annotations.Argument;
-import org.ballerinalang.natives.annotations.BallerinaFunction;
-import org.ballerinalang.natives.annotations.ReturnType;
-
-/**
- * Native function ballerina.model.system:getEnv.
- *
- * @since 0.8.0
- */
-@BallerinaFunction(
-        packageName = "ballerina.builtin",
-        functionName = "getEnv",
-        args = {@Argument(name = "key", type = TypeKind.STRING)},
-        returnType = {@ReturnType(type = TypeKind.STRING)},
-        isPublic = true
-)
-public class GetEnv extends AbstractNativeFunction {
-
-    public BValue[] execute(Context ctx) {
-        String varName = getStringArgument(ctx, 0);
-        if (varName != null && varName.trim().length() > 0) {
-            String varValue = System.getenv(varName);
-            if (varValue != null && varValue.trim().length() > 0) {
-                return getBValues(new BString(varValue));
-            }
-
-            return VOID_RETURN;
-        }
-
-        return VOID_RETURN;
-    }
-}
