+++ tests/ballerina-test/src/test/java/org/ballerinalang/test/debugger/VMDebuggerTest.java
@@ -185,10 +185,10 @@ public class VMDebuggerTest {
     @Test(description = "Testing debug paths in package init")
     public void testDebuggingPackageInit() {
         BreakPointDTO[] breakPoints = VMDebuggerUtil.createBreakNodeLocations(".",
-                "test-package-init.bal", 1, 7);
+                "test-package-init.bal", 3, 7);
 
         WorkerResults mainWorker = new WorkerResults(VMDebuggerUtil.createWorkerBreakPoints(".",
-                "test-package-init.bal", 1, 3, 12, 13, 14, 18, 12, 13, 14, 18, 12, 20, 7),
+                "test-package-init.bal", 3, 5, 14, 15, 16, 20, 14, 15, 16, 20, 14, 22, 9),
                 STEP_OVER, STEP_IN, STEP_OVER, STEP_OVER, STEP_OVER, STEP_OVER, STEP_OVER, STEP_OVER, STEP_OVER,
                 STEP_OVER, STEP_OVER, RESUME, RESUME);
         ExpectedResults expRes = new ExpectedResults(mainWorker, false);
