+++ tests/ballerina-test/src/test/java/org/ballerinalang/test/statements/foreach/ForeachNegativeTests.java
@@ -34,19 +34,19 @@ public class ForeachNegativeTests {
     public void testSemanticErrors() {
         CompileResult compile = BCompileUtil.compile("test-src/statements/foreach/foreach-negative.bal");
         Assert.assertEquals(compile.getErrorCount(), 13);
-        BAssertUtil.validateError(compile, 0, "too many variables are defined for iterable type 'string[]'", 3, 24);
-        BAssertUtil.validateError(compile, 1, "redeclared symbol 'i'", 12, 13);
-        BAssertUtil.validateError(compile, 2, "redeclared symbol 's'", 12, 16);
-        BAssertUtil.validateError(compile, 3, "undefined symbol 'i'", 22, 13);
-        BAssertUtil.validateError(compile, 4, "incompatible types: 'string' is not an iterable collection", 27, 18);
-        BAssertUtil.validateError(compile, 5, "invalid assignment in variable 'p.id'", 40, 13);
-        BAssertUtil.validateError(compile, 6, "too many variables are defined for iterable type 'string[]'", 48, 24);
-        BAssertUtil.validateError(compile, 7, "incompatible types: expected 'int', found 'string'", 54, 18);
-        BAssertUtil.validateError(compile, 8, "incompatible types: expected 'int', found 'string'", 54, 23);
+        BAssertUtil.validateError(compile, 0, "too many variables are defined for iterable type 'string[]'", 4, 24);
+        BAssertUtil.validateError(compile, 1, "redeclared symbol 'i'", 13, 13);
+        BAssertUtil.validateError(compile, 2, "redeclared symbol 's'", 13, 16);
+        BAssertUtil.validateError(compile, 3, "undefined symbol 'i'", 23, 16);
+        BAssertUtil.validateError(compile, 4, "incompatible types: 'string' is not an iterable collection", 28, 18);
+        BAssertUtil.validateError(compile, 5, "invalid assignment in variable 'p.id'", 41, 13);
+        BAssertUtil.validateError(compile, 6, "too many variables are defined for iterable type 'string[]'", 49, 24);
+        BAssertUtil.validateError(compile, 7, "incompatible types: expected 'int', found 'string'", 55, 18);
+        BAssertUtil.validateError(compile, 8, "incompatible types: expected 'int', found 'string'", 55, 23);
         BAssertUtil.validateError(compile, 9,
-                "incompatible types: 'json' cannot be convert to 'json[]', use cast expression", 61, 15);
-        BAssertUtil.validateError(compile, 10, "unreachable code", 73, 9);
-        BAssertUtil.validateError(compile, 11, "unreachable code", 78, 9);
-        BAssertUtil.validateError(compile, 12, "next cannot be used outside of a loop", 80, 5);
+                "incompatible types: 'json' cannot be convert to 'json[]', use cast expression", 62, 15);
+        BAssertUtil.validateError(compile, 10, "unreachable code", 74, 9);
+        BAssertUtil.validateError(compile, 11, "unreachable code", 79, 9);
+        BAssertUtil.validateError(compile, 12, "next cannot be used outside of a loop", 81, 5);
     }
 }
