+++ tests/ballerina-test/src/test/resources/test-src/debugger/test-package-init.bal
@@ -1,10 +1,12 @@
+import ballerina.io;
+
 const int val1 = calculateExp5(15, 10);
 
 int val2 = calculateExp5(20, 19);
 
 function main(string[] args) {
     int cal = val1 + val2;
-    println("package init - " + cal);
+    io:println("package init - " + cal);
 }
 
 function calculateExp5(int x, int y) (int) {
