+++ tests/ballerina-test/src/test/resources/test-src/expressions/unaryoperations/typeof-operation.bal
@@ -1,3 +1,4 @@
+import ballerina.io;
 function refTypeAccessTestTrivialEqualityPositiveCase() (int) {
     int temp_int = 2;
     int temp_int_1 = 5;
@@ -430,14 +431,14 @@ function testCheckTypeOfJson() (json, json[], string, int, float, boolean){
     if (typeof j1 == typeof json) {
         j = j1;
     }
-    
-    println(typeof j2);
-    println(typeof json[]);
+
+    io:println(typeof j2);
+    io:println(typeof json[]);
     if (typeof j2 == typeof json[]) {
         var ja, e = (json[]) j2;
-        println(e);
+        io:println(e);
     } else {
-        println("error on json[] cast");
+        io:println("error on json[] cast");
     }
 
     if (typeof j3 == typeof string) {
