+++ tests/ballerina-test/src/test/resources/test-src/nativeimpl/functions/system-test.bal
@@ -1,42 +1,43 @@
+import ballerina.io;
 function testPrintAndPrintlnString(string s1, string s2){
-    println(s1);
-    print(s2);
+    io:println(s1);
+    io:print(s2);
     // output is equal to s1\ns2
 }
 
 function testPrintAndPrintlnInt(int v1, int v2){
-    println(v1);
-    print(v2);
+    io:println(v1);
+    io:print(v2);
     // output is equal to v1\nv2
 }
 
 function testPrintAndPrintlnFloat(float v1, float v2){
-    println(v1);
-    print(v2);
+    io:println(v1);
+    io:print(v2);
     // output is equal to v1\nv2
 }
 
 function testPrintAndPrintlnBoolean(boolean v1, boolean v2){
-    println(v1);
-    print(v2);
+    io:println(v1);
+    io:print(v2);
     // output is equal to v1\nv2
 }
 
 function testPrintAndPrintlnConnector() {
     Foo f1 = create Foo();
     Foo f2 = create Foo();
-    println(f1);
-    print(f2);
+    io:println(f1);
+    io:print(f2);
 }
 
 function testPrintAndPrintlnFunctionPointer() {
     function (int, int) returns (int) addFunction = func1;
-    println(addFunction);
-    print(addFunction);
+    io:println(addFunction);
+    io:print(addFunction);
 }
 
 function printNewline() {
-    print("hello\n");
+    io:print("hello\n");
 }
 
 function testSleep(int timeoutv) {
