+++ tests/ballerina-test/src/test/resources/test-src/net/ws/http-to-websocket-upgrade-without-upgrade-ann.bal
@@ -1,3 +1,4 @@
+import ballerina.io;
 import ballerina.net.http;
 import ballerina.net.ws;
 
@@ -18,7 +19,7 @@ service<http> httpService {
     resource testResource(http:Connection conn, http:InRequest req) {
         http:OutResponse resp = {};
         string payload = req.getStringPayload();
-        println(payload);
+        io:println(payload);
         resp.setStringPayload("I received");
         _ = conn.respond(resp);
     }
@@ -27,11 +28,11 @@ service<http> httpService {
 service<ws> wsService  {
 
     resource onOpen(ws:Connection conn) {
-        println("New WebSocket connection: " + conn.getID());
+        io:println("New WebSocket connection: " + conn.getID());
     }
 
     resource onTextMessage(ws:Connection conn, ws:TextFrame frame) {
-        println(frame.text);
+        io:println(frame.text);
         conn.pushText(frame.text);
     }
 }
