+++ tests/ballerina-test/src/test/resources/test-src/services/service-negative.bal
@@ -1,29 +1,29 @@
+import ballerina.io;
 import ballerina.net.http;
-
 service<http> FooService {
 
     resource test1 (string dummyParam) {
-        println("test1-before");
+        io:println("test1-before");
         break;
-        println("test1-after");
+        io:println("test1-after");
     }
 
     resource test2 (string dummyParam) {
-        println("test2-before");
+        io:println("test2-before");
         next;
-        println("test2-after");
+        io:println("test2-after");
     }
 
     resource test3 (string dummyParam) {
-        println("test3-before");
+        io:println("test3-before");
         abort;
-        println("test3-after");
+        io:println("test3-after");
     }
 
     resource test4 (string dummyParam) {
-        println("test4-before");
+        io:println("test4-before");
         return;
-        println("test4-after");
+        io:println("test4-after");
     }
 
     resource test5 (string dummyParam) {
