+++ tests/ballerina-test/src/test/resources/test-src/workers/basic-worker-actions.bal
@@ -1,3 +1,4 @@
+import ballerina.io;
 function workerDeclTest() {
    worker wx {
      int a = 20;
@@ -9,7 +10,7 @@ function workerDeclTest() {
 	     int y = 0;
 	     int g = y + 1;
 	   }
-	} join (all) (map results) { println(results); }
+	} join (all) (map results) { io:println(results); }
    }
    worker wy { }
 }
@@ -29,7 +30,7 @@ function forkJoinWithMessageParsingTest() (int) {
 	     a <- w1;
 	     b -> w1;
 	   }
-	} join (all) (map results) { println(results); }
+	} join (all) (map results) { io:println(results); }
 	return x;
 }
 
@@ -50,7 +51,7 @@ function forkJoinWithSingleForkMessages() (int) {
 	     b -> w1;
 	     b -> fork;
 	   }
-	} join (all) (map results) { println(results); }
+	} join (all) (map results) { io:println(results); }
 	return x;
 }
 
@@ -86,7 +87,7 @@ function forkJoinWithMultipleForkMessages() (int) {
 	     b -> w1;
 	     a, b -> fork;
 	   }
-	} join (all) (map results) {  println(results);  }
+	} join (all) (map results) {  io:println(results);  }
 	return x;
 }
 
@@ -125,7 +126,7 @@ function forkJoinWithSomeJoin() (map) {
 	     int b = 15;
 	     m["x"] = b;
 	   }
-	} join (some 1) (map results) {  println(results);  }
+	} join (some 1) (map results) {  io:println(results);  }
 	return m;
 }
