+++ tests/ballerina-test/src/test/resources/test-src/workers/not-so-basic-worker-actions.bal
@@ -1,3 +1,4 @@
+import ballerina.io;
 import ballerina.net.http;
 
 function forkJoinWithTimeoutTest1() (map) {
@@ -106,7 +107,7 @@ function forkJoinWithSomeSelectedJoin1() (map) {
          sleep(1000);
 	     m["x"] = b;
 	   }
-	} join (some 1 w2, w3) (map results) {  println(results);  }
+	} join (some 1 w2, w3) (map results) {  io:println(results);  }
 	return m;
 }
 
@@ -134,7 +135,7 @@ function forkJoinWithSomeSelectedJoin2() (map) {
 	     (a * 2) -> w1;
 	     a <- w2;
 	   }
-	} join (some 1 w1, w2, w3) (map results) {  println(results);  }
+	} join (some 1 w1, w2, w3) (map results) {  io:println(results);  }
 	return m;
 }
 
@@ -164,7 +165,7 @@ function forkJoinWithSomeSelectedJoin3() (map) {
 	     (a * 2) -> w1;
 	     m["x"] <- w2;
 	   }
-	} join (some 1 w2, w3) (map results) {  println(results);  }
+	} join (some 1 w2, w3) (map results) {  io:println(results);  }
 	return m;
 }
 
@@ -187,7 +188,7 @@ function forkJoinWithSomeSelectedJoin4() (map) {
 	     sleep(1000);
 	     m["x"] = a * 2;
 	   }
-	} join (some 2 w1, w2, w3) (map results) {  println(results);  }
+	} join (some 2 w1, w2, w3) (map results) {  io:println(results);  }
 	return m;
 }
 
@@ -243,7 +244,7 @@ function forkJoinWithAllSelectedJoin1() (map) {
 	     (a * 2) -> w1;
 	     m["x"] <- w2;
 	   }
-	} join (all w2, w3) (map results) {  println(results);  }
+	} join (all w2, w3) (map results) {  io:println(results);  }
 	return m;
 }
 
@@ -413,12 +414,12 @@ function forkJoinWithStruct () (string result) {
     fork {
         worker w1 {
             foo f = {x:1, y:"w1"};
-            println(f);
+            io:println(f);
             f -> fork;
         }
         worker w2 {
             float f = 10.344;
-            println("[w2] f: " + f);
+            io:println("[w2] f: " + f);
             f -> fork;
         }
     } join (all) (map results) {
@@ -449,7 +450,7 @@ function forkJoinWithSameWorkerContent () (string result) {
         }
 
     } join (all) (map results) {
-        println(results);
+        io:println(results);
     }
     fork {
         worker w1 {
@@ -479,6 +480,6 @@ function testWorkerStackCreation ()(int) {
         return 1;
     }
     worker w2 {
-        println("testWorkerStackCreation done.");
+        io:println("testWorkerStackCreation done.");
     }
 }
